+++ tests/testthat/test-fork.R
@@ -15,11 +15,17 @@ with_mock_HTTP({
             '{"element":"shoji:entity",',
             '"body":{"dataset":"api/datasets/3/",',
             '"autorollback":false,"force":false}}')
-        expect_POST(mergeFork(ds1, ds2, force=TRUE),
+        expect_POST(with(consent(), mergeFork(ds1, ds2, force=TRUE)),
             "api/datasets/1/actions/",
             '{"element":"shoji:entity",',
             '"body":{"dataset":"api/datasets/3/",',
             '"autorollback":true,"force":true}}')
+    })
+
+
+    test_that('"Force" merge requires confirmation', {
+        expect_error(mergeFork(ds1, ds2, force=TRUE),
+            "Must confirm force merge")
     })
 })
 
@@ -69,8 +75,10 @@ with_test_authentication({
         expect_true(is.published(f3))
     })
 
-    delete(f2)
-    delete(f3)
+    with_consent({
+        delete(f2)
+        delete(f3)
+    })
     test_that("If you delete a fork, it disappears from upstream forks catalog", {
         expect_identical(names(refresh(forks(ds))), f1.name)
     })
@@ -109,7 +117,7 @@ with_test_authentication({
         expect_identical(name(f1$v1), v1copy$name)
         expect_identical(alias(f1$v1), v1copy$alias)
     })
-    f1$v1 <- NULL
+    with_consent(f1$v1 <- NULL)
     f1$v1 <- v1copy
 
     ## Assert those things
@@ -152,7 +160,7 @@ with_test_authentication({
         child <- forkDataset(parent)
 
         names(categories(parent$v4))[1:2] <- c("d", "e")
-        parent$v5 <- NULL
+        with_consent(parent$v5 <- NULL)
 
         test_that("Before merging from parent", {
             expect_identical(names(categories(parent$v4))[1:2], c("d", "e"))